+++ test/unit/spec/signaling/v2/twilioconnectiontransport.js
@@ -33,7 +33,7 @@ describe('TwilioConnectionTransport', () => {
       [{ video: {} }, { video: {} }],
       [{ video: { mode: 'foo' } }, { video: { mode: 'foo' } }],
       // eslint-disable-next-line
-      [{ video: { maxSubscriptionBitrate: 2048 } }, { video: { max_subscription_bandwidth: 2048 } }],
+      [{ video: { maxSubscriptionBitrate: 2048 } }, { video: { max_subscription_bandwidth: 2048000 } }],
       // eslint-disable-next-line
       [{ video: { maxSwitchedOnTracks: 2 } }, { video: { max_tracks: 2 } }],
       // eslint-disable-next-line
@@ -64,7 +64,7 @@ describe('TwilioConnectionTransport', () => {
             // eslint-disable-next-line
             active_speaker_priority: 'baz',
             // eslint-disable-next-line
-            max_subscription_bandwidth: 4096,
+            max_subscription_bandwidth: 4096000,
             // eslint-disable-next-line
             max_tracks: 5,
             mode: 'bar'