+++ CHANGELOG.md
@@ -2,8 +2,8 @@ The Twilio Programmable Video SDKs use [Semantic Versioning](http://www.semver.o
 
 **Version 1.x reached End of Life on September 8th, 2021.** See the changelog entry [here](https://www.twilio.com/changelog/end-of-life-complete-for-unsupported-versions-of-the-programmable-video-sdk). Support for the 1.x version ended on December 4th, 2020.
 
-3.0.0-preview.2 (in progress)
-=============================
+3.0.0-preview.2 (August 18, 2022)
+=================================
 
 Changes
 -------